Struct hyper_socks::Socks4HttpsConnector

pub struct Socks4HttpsConnector<S> {
    // some fields omitted
}

A connector that will produce HttpsStreams proxied over a SOCKS4 server.

Methods

impl<S: SslClient> Socks4HttpsConnector<S>

fn new<T: ToSocketAddrs>(proxy: T, userid: &str, ssl: S) -> Result<Self>

Creates a new Socks4HttpsConnector which will connect to the specified proxy with the specified userid, and use the provided SSL implementation to encrypt the resulting stream.